# mcp-server/openapi.yaml
openapi: "3.1.0"
info:
  title: RASS (Retrieval-Augmented Semantic Search) API
  version: "1.0.0"
  description: |
    Production-grade multi-service RAG system with document ingestion, hybrid retrieval,
    SSE streaming, multi-tenant workspaces, RBAC, audit logging, and API key authentication.

    ## Authentication
    Most endpoints require a Bearer JWT token from POST /api/auth/login.
    Machine-to-machine clients may use an API key via the Authorization header
    with the format `ApiKey <raw_key>`.

    ## SSE Streaming
    POST /api/stream-ask returns Server-Sent Events. See docs/api/streaming.md.
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    KnowledgeBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: string
          format: uuid
        openSearchIndex:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/knowledge-bases:
    post:
      tags: [Knowledge Bases]
      summary: Create a knowledge base (provisions OpenSearch index)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "Medical Guidelines KB"
                description:
                  type: string
                  maxLength: 500
              required: [name]
      responses:
        "201":
          description: Knowledge base created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
    get:
      tags: [Knowledge Bases]
      summary: List accessible knowledge bases
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Knowledge bases
          content:
            application/json:
              schema:
                type: object
                properties:
                  knowledgeBases:
                    type: array
                    items:
                      $ref: "#/components/schemas/KnowledgeBase"
  /api/knowledge-bases/{kbId}:
    get:
      tags: [Knowledge Bases]
      summary: Get knowledge base details
      security:
        - BearerAuth: []
      parameters:
        - name: kbId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "404":
          description: Not found
    delete:
      tags: [Knowledge Bases]
      summary: Delete knowledge base and all its documents
      security:
        - BearerAuth: []
      parameters:
        - name: kbId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
        "403":
          description: Only the owner can delete
  /api/knowledge-bases/{kbId}/members:
    post:
      tags: [Knowledge Bases]
      summary: Add or update a member (owner only)
      security:
        - BearerAuth: []
      parameters:
        - name: kbId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum: [OWNER, EDITOR, VIEWER]
              required: [userId]
      responses:
        "201":
          description: Member added
...